- Refactoring web service for modular communication - Not currently runnable
